requirements.txt
discord.py==2.3.2
yt-dlp==2025.10.22
flask==3.1.2
